backend/requirements.txt
pymongo[srv]==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
sentence-transformers
chromadb
python-dateutil
groq
